requirements.txt
astroid==2.12.14
attrs==22.2.0
beautifulsoup4==4.11.1
black==22.12.0
click==8.1.3
dill==0.3.6
django-htmlmin==0.11.0
exceptiongroup==1.1.0
Flask==2.2.2
gunicorn==20.1.0
html5lib==1.1
importlib-metadata==6.0.0
iniconfig==1.1.1
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
pylint==2.15.9
pytest==7.2.0
pytoolconfig==1.2.4
rope==1.6.0
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.12.0
